squash! fix: share home and xdg dirs for file
Make `podman machine` commands and the associated temporary dir setup
conditional on macOS only where the VM is needed.

Running these on Linux forces rootless mode which requires some
additional dependencies and is presumably less efficient:

    not ok 1 setup_file failed
    # (from function `start_podman_machine' in file containerize.bats, line 134,
    #  from function `setup_file' in test file containerize.bats, line 192)
    #   `start_podman_machine' failed with status 125
    # Looking up Podman Machine image at quay.io/podman/machine-os:5.2 to create VM
    # Getting image source signatures
    # Copying blob sha256:6d2791f7141a330267b6d5c83b7a5a974a66b9a05a99ab2b1d5e1932c18ed68d
    # Copying config sha256:44136fa355b3678a1146ad16f7e8649e94fb4fc21fe77e8310c060f61caaff8a
    # Writing manifest to image destination
    # 6d2791f7141a330267b6d5c83b7a5a974a66b9a05a99ab2b1d5e1932c18ed68d
    # Error: open /etc/timezone: no such file or directory
    # Error: command required for rootless mode with multiple IDs: exec: "newuidmap": executable file not found in $PATH
    # Error: podman-machine-default: VM does not exist

From docs: https://docs.podman.io/en/latest/markdown/podman-machine-init.1.html

> Rootless only.
>
> Podman on MacOS and Windows requires a virtual machine. This is because containers are Linux - containers do not run on any other OS because containers’ core functionality are tied to the Linux kernel. Podman machine must be used to manage MacOS and Windows machines, but can be optionally used on Linux.

+++ cli/tests/containerize.bats
@@ -186,10 +186,14 @@ setup_file() {
   # As a side effect the individual tests will run faster
   # because podman does not need to serialize writes to the cache.
   export BATS_NO_PARALLELIZE_WITHIN_FILE=true
-  podman_global_dirs_setup
-  common_test_setup
-  podman_home_setup
-  start_podman_machine
+
+  # Only for macOS, don't force rootless on Linux.
+  if ! is_linux; then
+    podman_global_dirs_setup
+    common_test_setup
+    podman_home_setup
+    start_podman_machine
+  fi
 }
 
 teardown() {
@@ -200,9 +204,12 @@ teardown() {
 teardown_file() {
   podman_cache_reset
   common_file_teardown
-  podman machine stop
-  rm -rf "$SHORT_TMP"
-  rm -rf "$FLOX_TEST_HOME"
+
+  if ! is_linux; then
+    podman machine stop
+    rm -rf "$SHORT_TMP"
+    rm -rf "$FLOX_TEST_HOME"
+  fi
 }
 
 # ---------------------------------------------------------------------------- #
